requirements-dev.txt
alabaster==0.7.10
amqp==2.1.4
appdirs==1.4.3
babel==2.3.4
beautifulsoup4==4.5.3
billiard==3.5.0.2
celery==4.0.2
click==6.7
coverage==4.3.4
django-webtest==1.9.1
docutils==0.13.1
first==2.0.1
graphviz==0.6
imagesize==0.7.1
Jinja2==2.9.5
kombu==4.0.2
MarkupSafe==1.0
packaging==16.8
pip-tools==1.8.0
pluggy==0.4.0
py==1.4.32
pyenchant==1.6.8
Pygments==2.2.0
pyparsing==2.2.0
pytest-django==3.1.2
pytest==3.0.6
pytz==2016.10
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
sphinx==1.5.3
sphinxcontrib-spelling==2.3.0
tox==2.6.0
vine==1.1.3
virtualenv==15.1.0
waitress==1.0.2
WebOb==1.7.1
webtest==2.0.26

# The following packages are considered to be unsafe in a requirements file:
# setuptools
